+++ packages/extension/src/view/devtools/components/tabContent/cookies/cookiesListing/index.tsx
@@ -22,12 +22,10 @@ import { Resizable } from 're-resizable';
 /**
  * Internal dependencies.
  */
-import {
-  type CookieTableData,
-  useCookieStore,
-} from '../../../../stateProviders/syncCookieStore';
+import { useCookieStore } from '../../../../stateProviders/syncCookieStore';
 import CookieDetails from './cookieDetails';
 import CookieTable from './cookieTable';
+import { filterCookiesByFrame } from '../utils/filterCookiesByFrame';
 
 const CookiesListing = () => {
   const { cookies, selectedFrame, tabFrames } = useCookieStore(({ state }) => ({
@@ -36,21 +34,10 @@ const CookiesListing = () => {
     tabFrames: state.tabFrames,
   }));
 
-  const calculatedCookies = useMemo(() => {
-    const frameFilteredCookies: { [key: string]: CookieTableData } = {};
-
-    if (cookies && selectedFrame && tabFrames && tabFrames[selectedFrame]) {
-      Object.entries(cookies).forEach(([key, cookie]) => {
-        tabFrames[selectedFrame].frameIds?.forEach((frameId) => {
-          if (cookie.frameIdList?.includes(frameId)) {
-            frameFilteredCookies[key] = cookie;
-          }
-        });
-      });
-    }
-
-    return Object.values(frameFilteredCookies);
-  }, [cookies, selectedFrame, tabFrames]);
+  const filteredCookies = useMemo(
+    () => filterCookiesByFrame(cookies, tabFrames, selectedFrame),
+    [cookies, selectedFrame, tabFrames]
+  );
 
   return (
     <div className="h-full flex flex-col">
@@ -68,7 +55,7 @@ const CookiesListing = () => {
           left: false,
         }}
       >
-        <CookieTable cookies={calculatedCookies} />
+        <CookieTable cookies={filteredCookies} />
       </Resizable>
       <div className="w-full h-full bg-white border-2 border-gray-300 shadow overflow-auto">
         <CookieDetails />
